requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.28.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
python-crontab>=2.6.0
python-telegram-bot>=20.0
